Remove cross-compiled package from workspace again
First of all, it stops "cargo test" from working, as it tries to execute
the ARM binary of "hello" in unit-test mode.

More seriously, it requires nightly Cargo, meaning all users of the
whole of Cotton would also need to use nightly Cargo.

The first alternative tried was to use build.rs to run a second, inner
Cargo. This started off promisingly, but deadlocked (not entirely
surprisingly) when building the cotton-stm32-eth crate itself. (What would
it do -- *re*-run build.rs?)

So the solution adopted is to move the actual code of cotton-stm32-eth into
a sub-package, cotton-stm32-eth-core -- so that when the cross-compiled
package references it, there's no recursion.

This isn't quite ideal, because you don't get progress reports from the
outer Cargo while running the inner Cargo -- but at least everything gets
correctly built.

We additionally need to pass --target-dir to the inner Cargo, otherwise
it deadlocks locking the *output* directory.