Revert pinning of transitive dependencies
A better way of dealing with MSRV-breaking changes in transitive
dependencies, is for CI to do its "cargo fetch" with MSRV, so it just
doesn't update those dependencies in the first place.

[Edit:] Wait, "cargo fetch" just doesn't work like that -- it will
cheerfully fetch so-called "semver-compatible" updates that have too-new
MSRVs. For now, punt this whole issue, as cotton will soon be moving to
new MSRV 1.65, as that's what all the good smoltcp's need.