Don't assume indexes are dense
The previous code assumed that the kernel's network interface indexes,
were always 1,2,3... with no gaps. This doesn't seem to be guaranteed
(I can't find either a guarantee, or a case that causes a gap), so for
robustness's sake, the code now checks, using if_nametoindex, that it's
getting the interface indexes correct.