rp2040: Clear DBGPAUSE bits on timer peripheral
RP2040 system-tests were intermittently failing. It turned out that
timer.delay_ms() was hanging, because the timer wasn't incrementing.

The timer doesn't increment if either RP2040 core is under debug, unless
the DBGPAUSE bits are cleared, which they aren't by default. Clearing
those bits makes the tests pass.

There is no neat and tidy method on hal::Timer to clear these bits, and
they can't be cleared before hal::Timer::new because it resets the
peripheral. So we have to steal the peripheral, but that's OK because
we only access the DBGPAUSE register, which nobody else is accessing.

It is not clear why the test failures were intermittent -- perhaps the
debugger only intermittently successfully connects to the second core?

https://github.com/raspberrypi/debugprobe/issues/45 is not this exact
problem but led me to the solution.