Update minimum version of futures-util
This is a complicated one. There was a soundness bug in futures-task
https://github.com/rust-lang/futures-rs/issues/2795 so it was fixed in a
point release (0.3.31) even though it's a breaking change.

But other crates, including futures-util, weren't updated to "depend on"
the newer release, even though they now don't compile against the older
release.

This suddenly broke the cotton-minver build, which used the old
futures-util 0.3.11 (because it *directly* depends on it) but the new
futures-task 0.3.31 (because it only *indirectly* depends on it, and the
minver build does direct-minimal-versions only).

Fix this by depending on a later version of futures-util.