usb: Use ConstStaticCell for UsbStatics
On the one hand, it's a bit cheat-y to use ConstStaticCell just because
UsbStatics is !Sync... ConstStaticCell<T> is Sync for all T, whereas an
actual static variable is only allowed for T: Sync.

On the other hand, Box::leak produces a &'static T for any T, so this
is no worse than that.

(Answering my own question:) Both the ConstStaticCell and Box::leak
cases are fine because *only one thread of execution* gets the &'static.