usb: Tweak control-transfer state machine
The zero-length-packet (ZLP) rules are quite subtle.

For bulk transfers, a transfer that's an exact multiple of the packet size
needs a ZLP at the end, to indicate that the transfer is complete.

The data phase of a control transfer does NOT need a ZLP, because the
transfer length was signalled in advance in the setup packet -- indeed
it mustn't have one, and the Auvipal hub stalls if it sees one. The
exception is if the data phase of a control IN transfer is shorter than
requested AND is a multiple of the packet size. (But as the host stack,
that's not our problem.)

However the status phase of a control transfer, which is only ever of
length zero, must still consist of a packet!

The packetiser/depacketiser code will need a bit of revisiting to support
bulk transfers.