USB: Fix race condition in RP2040 USB IRQ handler
The previous code handled all then-active interrupts, then cleared all
then-active interrupts. But it did so by determining "then-active
interrupts" *twice*, which loses an interrupt if it happens in-between
those two checks.

This is relatively uncommon at high optimisation levels (because the
window for the interrupt is so short), but was exposed when compiling
with lower optimisation levels.

Many thanks to thejpster on Github for the test case that ended up
exposing this bug!